TASK: Expose `getInlinedDataFromBackend` invoke from outside
Also move appContainer
          csrfToken
          systemEnv back to the Ui as no good abstraction was found.
