+++ packages/neos-ui-configuration/src/bootstrap/index.ts
@@ -1,0 +1,1 @@
+export {getInlinedDataFromBackend} from './getInlinedDataFromBackend'
