+++ packages/neos-ui-configuration/src/systemEnv.ts
@@ -1,5 +1,8 @@
-import {systemEnv} from './system';
-
 export function isDevelopmentContext(): boolean {
+    const systemEnv = document.getElementById('appContainer')?.dataset.env;
+    if (!systemEnv) {
+        // should not happen
+        return true;
+    }
     return systemEnv === 'Development' || systemEnv.startsWith('Development/')
 }
